Add cfg option gnu_time64_abi
The gnu_time64_abi option indicates that a 32-bit arch should use
64-bit time_t and 64-bit off_t and simliar file related types.

32-bit platforms are identified by CARGO_CFG_TARGET_POINTER_WIDTH, but
x86_64-unknown-linux-gnux32 is a 64-bit platform and should not use
gnu_time64_abi even if it has 32-bit pointers.

riscv32 is a 32-bit platform but has always used 64-bit types for
time and file operations.  It should not use the gnu_time64_abi

The default - all relevant platforms should use 64-bit time - can be
overridden by setting the RUST_LIBC_TIME_BITS environment variable to
32.  It can also be set to 64 to force gnu_time64_abi, or default
which is the same as leaving it unset.  The last two options are
mainly useful for CI flows.